+++ Service/Interfaces/IAdminService.cs
@@ -5,6 +5,6 @@ namespace Service.Interfaces
 	[ScopedService]
 	public interface IAdminService
 	{
-		void UpdateUsername(string name, int id, int userId);
+		void UpdateUsername(string name, long id, long userId);
 	}
 }
